move document components into appropriate directory
